tools/requirements_zmq.txt
pyzmq>=25.0
pyserial>=3.5
